+++ src/olmo_core/launch/beaker.py
@@ -247,15 +247,17 @@ class BeakerLaunchConfig(Config):
     Uses hostname constraints to restrict the hostnames on which the experiment runs. This is currently
     only supported for Augusta clusters, and can benefit performance by forcing the use of colocated nodes.
 
-    This is NOT recommended to be used lower priority preemptible jobs, since hostname constraints are not
+    This is NOT recommended to be used with lower priority preemptible jobs, since hostname constraints are not
     updated on preemption.
     """
 
     num_execution_units: Optional[int] = None
     """
-    Number of \"execution units\", defaults to ``max(1, num_nodes // 32)``. An \"execution unit\" is abstraction
+    Number of "execution units", defaults to ``max(1, num_nodes // 32)``. An "execution unit" is abstraction
     for any node-using entity of which 1 or more copies are run, where each unit wants its nodes to be
     from colocated hardware (e.g., a model replica for large jobs, or a full distributed model for small jobs).
+
+    For internal experiments, this defaults to the number of data-parallel model replicas instead.
     """
 
     # NOTE: don't assign a type here because omegaconf can't validate arbitrary classes