+++ CHANGELOG.md
@@ -7,6 +7,12 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## Unreleased
 
+### Fixed
+
+- Fixed parsing username+password git remote URLs in `launch.beaker` module.
+
+## [v2.2.0](https://github.com/allenai/OLMo-core/releases/tag/v2.2.0) - 2025-08-26
+
 ### Added
 
 - Added option to set LR scheduler based on tokens instead of steps (e.g. `--train_module.scheduler.units=tokens`).
@@ -37,6 +43,9 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 - Added `namespace` option to `nn.buffer_cache.BufferCache`.
 - Added the option to configure `head_stride` for context parallelism with ring-flash-attn.
 - Added the option to group multiple npy source files together for packing with the packed FSL dataset by setting `source_group_size` to an integer greater than 1.
+- Added optional hostname constraints for beaker experiments on Google clusters.
+- Added `load_optim_state: Optional[bool]` option to `Trainer.load_checkpoint()`.
+- Added `GenerationModule` for OLMo-core native autoregressive generation with support for kv caching.
 - Added optional hostname constraints for beaker experiments on Google clusters.
 
 ### Changed
@@ -75,6 +84,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 - Fixed MFU computation to work with FSDP, corrected some device specs.
 - Optimization: avoid redundant calls to `model.train()` in `TransformerTrainModule`.
 - `NumpyDatasetConfig.expand_glob` now works with remote directories.
+- Fixed Attention block sharding when TP and head-wise QK norm are both applied.
 
 ## [v2.1.0](https://github.com/allenai/OLMo-core/releases/tag/v2.1.0) - 2025-04-14